fix(ci): python-full-coverage — split oversized verdict-retry part_001 test file
Move three hosted provider-failure rollback tests into part_004 so
test_pr_monitor_verdict_retry_part_001.py stays under the 1500-line limit.
